+++ application/components/LCR/src/LCR_X.cpp
@@ -38,12 +38,12 @@ void LCR_X::toGui() const
     IL::getLCR_Hub().toGui(mId, mStateToGui, getUbkToGui());
 }
 
-void LCR_X::toFld(const INT32 state) const
+void LCR_X::toFld(const UINT8 state) const
 {
     IL::getLCR_Hub().toFld(mId, state);
 }
 
-void LCR_X::fromGui(const INT32 state)
+void LCR_X::fromGui(const UINT8 state)
 {
     switch (state)
     {
@@ -59,7 +59,7 @@ void LCR_X::fromGui(const INT32 state)
     }
 }
 
-bool LCR_X::validState(const INT32 state)
+bool LCR_X::validState(const UINT8 state)
 {
     bool ok = false;
     switch (state)
@@ -77,7 +77,7 @@ bool LCR_X::validState(const INT32 state)
     return ok;
 }
 
-bool LCR_X::validUbk(const INT32 state)
+bool LCR_X::validUbk(const UINT8 state)
 {
     bool ok = false;
     switch (state)
@@ -95,7 +95,7 @@ bool LCR_X::validUbk(const INT32 state)
     return ok;
 }
 
-void LCR::fromFld(const INT32 state, const INT32 ubk)
+void LCR::fromFld(const UINT8 state, const UINT8 ubk)
 {
     if (state == mStateToGui)
     { pass(); }
@@ -108,7 +108,7 @@ void LCR::fromFld(const INT32 state, const INT32 ubk)
     }
 }
 
-void LCR_UBK::fromFld(const INT32 state, const INT32 ubk)
+void LCR_UBK::fromFld(const UINT8 state, const UINT8 ubk)
 {
     if (
         (state == mStateToGui) and
