+++ application/components/TSW/TSW_Hub.h
@@ -16,8 +16,8 @@ public:
     inline TSW_Hub()
     {}
 
-    void toFld(size_t id, INT32 cmd) const;
-    void toGui(size_t id, INT32 state) const;
+    void toFld(size_t id, UINT8 cmd) const;
+    void toGui(size_t id, UINT8 state) const;
 
     void fromDsp(size_t pos, const FldState& tele) const;
     void fromDsp(size_t pos, const GuiCmd&   tele) const;
