+++ specification/ifs/I_TSW_Hub.h
@@ -12,8 +12,8 @@
 class I_TSW_Hub
 {
 public:
-    virtual void toFld(size_t id, INT32 cmd) const = 0;
-    virtual void toGui(size_t id, INT32 state) const = 0;
+    virtual void toFld(size_t id, UINT8 cmd) const = 0;
+    virtual void toGui(size_t id, UINT8 state) const = 0;
 
     virtual void fromDsp(size_t pos, const FldState& tele) const = 0;
     virtual void fromDsp(size_t pos, const GuiCmd&   tele) const = 0;
