+++ testing/testenv/mocks/M_SIG_Hub.h
@@ -17,20 +17,20 @@ namespace test
     public:
         MOCK_CON(SIG_Hub)
 
-        inline void toFld(size_t id, INT32 state, INT32 speed) const
+        inline void toFld(size_t id, UINT8 state, UINT8 speed) const
         {
             call("toFld").PARAM(id).PARAM(state).PARAM(speed);
         }
-        inline void expectToFld(size_t id, INT32 state, INT32 speed = 0) const
+        inline void expectToFld(size_t id, UINT8 state, UINT8 speed = 0) const
         {
             expect("toFld").PARAM(id).PARAM(state).PARAM(speed);
         }
 
-        inline void toGui(size_t id, INT32 state, INT32 speed) const
+        inline void toGui(size_t id, UINT8 state, UINT8 speed) const
         {
             call("toGui").PARAM(id).PARAM(state).PARAM(speed);
         }
-        inline void expectToGui(size_t id, INT32 state, INT32 speed = 0) const
+        inline void expectToGui(size_t id, UINT8 state, UINT8 speed = 0) const
         {
             expect("toGui").PARAM(id).PARAM(state).PARAM(speed);
         }
